pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentforge"
version = "0.1.0"
description = "Autonomous AI agent framework with verified execution"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "AgentForge Team" }
]
keywords = [
    "ai",
    "agents",
    "automation",
    "verification",
    "llm",
    "anthropic",
    "claude",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Quality Assurance",
    "Typing :: Typed",
]

dependencies = [
    # CLI and formatting
    "click>=8.1.0",
    "rich>=13.0.0",

    # Data handling
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # LLM integration
    "anthropic>=0.30.0",
    "httpx>=0.27.0",

    # Async support
    "aiofiles>=23.0.0",

    # Web dashboard (optional, but included for full experience)
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",

    # File watching and git
    "watchdog>=4.0.0",
    "gitpython>=3.1.0",

    # Environment and utilities
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "structlog>=24.0.0",

    # Code analysis (for refactoring)
    "rope>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "pre-commit>=3.6.0",
    "types-PyYAML",
    "types-aiofiles",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
agentforge = "agentforge.cli.main:cli"

[project.urls]
Homepage = "https://github.com/agentforge/agentforge"
Documentation = "https://agentforge.dev"
Repository = "https://github.com/agentforge/agentforge"
Issues = "https://github.com/agentforge/agentforge/issues"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentforge"]

# Include package data (default configs)
[tool.hatch.build.targets.wheel.force-include]
"contracts" = "agentforge/defaults/contracts"
"schemas" = "agentforge/defaults/schemas"

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument (common in click)
]

[tool.ruff.lint.isort]
known-first-party = ["agentforge"]

# MyPy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "anthropic.*",
    "rope.*",
    "yaml.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src/agentforge"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
